More work on k-comodules

+++ src/page.rs
@@ -4,7 +4,7 @@ pub struct Page {
     pub degrees: Vec<char>,
     pub x_formula: String,
     pub y_formula: String,
-    pub generators: Vec<(usize, usize, Vec<usize>, Option<String>)>,
+    pub generators: Vec<(usize, usize, Vec<i32>, Option<String>)>,
     pub structure_lines: Vec<((usize,usize), (usize,usize), usize, String)>,
     pub differentials: Vec<((usize,usize), (usize,usize))>
 }
